+++ components-starters/components-starter-web/src/main/java/com/zoe/framework/components/starter/web/response/webmvc/ReturnResponseHandler.java
@@ -1,5 +1,6 @@
 package com.zoe.framework.components.starter.web.response.webmvc;
 
+import com.zoe.framework.components.starter.common.util.AnnoFindUtil;
 import com.zoe.framework.components.starter.web.configuration.GlobalResponseProperties;
 import com.zoe.framework.components.starter.web.response.DisableGlobalResponse;
 import com.zoe.framework.components.starter.web.response.DisableTransResultTypeResponse;
@@ -45,35 +46,43 @@ public class ReturnResponseHandler implements HandlerMethodReturnValueHandler {
                                   MethodParameter returnType,
                                   ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest) throws Exception {
-        if (globalResponseProperties == null || globalResponseProperties.getEnable()) {
-            boolean validPath = false;
-            try {
-                // 原request
-                HttpServletRequest nativeRequest = webRequest.getNativeRequest(HttpServletRequest.class);
-                // 当前路径包含在NotTransResultList的集合中，不进行全局替换。
-                validPath = UrlFilterUtil.valid(ObjectUtil.isNull(globalResponseProperties) ? new ArrayList<>() : globalResponseProperties.getDisableGlobalResponseList()
-                        , nativeRequest.getServletPath());
-            } catch (Exception e) {
-                log.error("返回值处理异常：", e);
-            }
-            // 使用 全局返回值
-            DisableGlobalResponse disableGlobalResponseMethod = returnType.getMethod().getAnnotation(DisableGlobalResponse.class);
-            DisableGlobalResponse disableGlobalResponseClass = returnType.getMethod().getClass().getAnnotation(DisableGlobalResponse.class);
-            if (!validPath && disableGlobalResponseMethod == null && disableGlobalResponseClass == null) {
-                // 判断是否支持feign
-                String useFeign = webRequest.getHeader(FeignConst.USE_FEIGN_NAME);
-                if (FeignConst.USE_FEIGN_VALUE.equals(useFeign)
-                        && ObjectUtil.isNotNull(globalResponseProperties)
-                        && !globalResponseProperties.getSupportFeign()) {
-                    // 不支持feign调用返回值的封装
-                } else {
-                    DisableTransResultTypeResponse disableTransResultTypeResponseMethod = returnType.getMethod().getAnnotation(DisableTransResultTypeResponse.class);
-                    DisableTransResultTypeResponse disableTransResultTypeResponseClass = returnType.getMethod().getClass().getAnnotation(DisableTransResultTypeResponse.class);
-                    boolean transResultResponse = disableTransResultTypeResponseMethod == null && disableTransResultTypeResponseClass == null;
-                    returnValue = responseHandleChainPattern.execute(returnValue, transResultResponse);
-                }
-            }
+        if (globalResponseProperties != null && !globalResponseProperties.isEnable()) {
+            delegate.handleReturnValue(returnValue, returnType, mavContainer, webRequest);
+            return;
         }
+        boolean validPath = false;
+        try {
+            // 原request
+            HttpServletRequest nativeRequest = webRequest.getNativeRequest(HttpServletRequest.class);
+            // 当前路径包含在NotTransResultList的集合中，不进行全局替换。
+            validPath = UrlFilterUtil.valid(ObjectUtil.isNull(globalResponseProperties) ? new ArrayList<>() : globalResponseProperties.getDisableGlobalResponseList()
+                    , nativeRequest.getServletPath());
+        } catch (Exception e) {
+            log.error("返回值处理异常：", e);
+        }
+        if (validPath) {
+            delegate.handleReturnValue(returnValue, returnType, mavContainer, webRequest);
+            return;
+        }
+        boolean disableGlobalResponse = AnnoFindUtil.has(returnType.getMethod(), DisableGlobalResponse.class);
+        if (disableGlobalResponse) {
+            delegate.handleReturnValue(returnValue, returnType, mavContainer, webRequest);
+            return;
+        }
+        // 使用 全局返回值
+        // 判断是否支持feign
+        String useFeign = webRequest.getHeader(FeignConst.USE_FEIGN_NAME);
+        if (FeignConst.USE_FEIGN_VALUE.equals(useFeign)
+                && ObjectUtil.isNotNull(globalResponseProperties)
+                && !globalResponseProperties.isSupportFeign()) {
+            // 不支持feign调用返回值的封装
+            delegate.handleReturnValue(returnValue, returnType, mavContainer, webRequest);
+            return;
+        }
+        // 针对IResult返回类型，是否需要处理
+        boolean disableTransResultTypeResponse = AnnoFindUtil.has(returnType.getMethod(), DisableTransResultTypeResponse.class);
+        disableTransResultTypeResponse = disableTransResultTypeResponse || globalResponseProperties.isDisableGlobalTransResultTypeResponse();
+        returnValue = responseHandleChainPattern.execute(returnValue, disableTransResultTypeResponse);
         delegate.handleReturnValue(returnValue, returnType, mavContainer, webRequest);
     }
 }