+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * refactor: IResult优化
 * refactor: Exception优化
 * refactor: starter-annotation优化
+* refactor: starter-web优化
 
 ##### 1.2.3.20220414-RELEASE